requirements.txt
chromadb==0.5.0
colorama==0.4.6

langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langchainhub==0.1.21
langsmith==0.2.11
openai==1.59.7

bs4==0.0.2
tqdm==4.66.4
nltk==3.8.1

python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
tiktoken==0.7.0

streamlit==1.41.1
